Day 11

+++ src/main/java/com/skillSwap/skillswap/Entity/Credit.java
@@ -11,9 +11,9 @@ public class Credit {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 
-	private int amount;
+	private Integer amount;
 
 	@ManyToOne
 	@JoinColumn(name = "user_id")
